coastal_backend/requirements.txt
Django==5.2.5
djangorestframework==3.16.1
django-cors-headers==4.4.0
celery==5.5.3
redis==6.4.0
requests==2.32.3
python-decouple==3.8
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
joblib>=1.3.0
